[AML-237] First/last name transposed

+++ src/SFA.DAS.EmployerApprenticeshipsService.Database/Membership.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[Membership]
 (
-    [AccountId] INT NOT NULL, 
-    [UserId] INT NOT NULL, 
-    [RoleId] INT NOT NULL, 
+    [AccountId] BIGINT NOT NULL, 
+    [UserId] BIGINT NOT NULL, 
+    [RoleId] TINYINT NOT NULL, 
     CONSTRAINT [FK_Membership_Account] FOREIGN KEY (AccountId) REFERENCES [Account]([Id]), 
     CONSTRAINT [FK_Membership_User] FOREIGN KEY (UserId) REFERENCES [User]([Id]), 
     CONSTRAINT [FK_Membership_Role] FOREIGN KEY (RoleId) REFERENCES [Role]([Id]),
