+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/IAccountRepository.cs
@@ -4,6 +4,6 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Data
 {
     public interface IAccountRepository
     {
-        Task<int> CreateAccount(string userId, string employerNumber, string employerName, string employerRef);
+        Task<long> CreateAccount(string userId, string employerNumber, string employerName, string employerRef);
     }
 }
