+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/IMembershipRepository.cs
@@ -7,8 +7,8 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Data
         Task<TeamMember> Get(long accountId, string email);
         Task<Membership> Get(long userId, long accountId);
         Task Remove(long userId, long accountId);
-        Task ChangeRole(long userId, long accountId, int roleId);
+        Task ChangeRole(long userId, long accountId, short roleId);
         Task<MembershipView> GetCaller(long accountId, string externalUserId);
-        Task Create(long userId, long accountId, int roleId);
+        Task Create(long userId, long accountId, short roleId);
     }
 }
