+++ src/SFA.DAS.LevyAggregationProvider.Worker/Model/SourceData.cs
@@ -4,7 +4,7 @@ namespace SFA.DAS.LevyAggregationProvider.Worker.Model
 {
     public class SourceData
     {
-        public int AccountId { get; set; }  
+        public long AccountId { get; set; }  
 
         public List<SourceDataItem> Data { get; set; }
     }
